Conflicts:
	.gitignore
	Classes/AnimationFactory.cpp
	Classes/Character.cpp
	Classes/GameManager.cpp
	Classes/GameObject.cpp
	Classes/GameWorld.cpp
	Classes/MapCreator.cpp
	Classes/MapObject.cpp
	Classes/SkillFactory.cpp
	Classes/Util.h
	proj.ios/TinyZodiacs.xcodeproj/project.pbxproj

+++ Classes/Character.h
@@ -73,7 +73,7 @@ public:
     virtual void BeginContact(b2Contact *contact);
     virtual void EndContact(b2Contact *contact);
     
-    virtual void setGroup(int group);
+    virtual void setGroup(uint16 group);
 };
 
 #endif /* defined(__SampleCocosProject__Character__) */
